refactor: remove eslint-disable as many as possible
follow commit 26ebfa0f8a6a58b0d3e7730133ce4d955057eaad
