--- src/MarkBlocked/modules/util/.eslintrc
@@ -1,6 +1,0 @@
-{
-	"rules": {
-		"@typescript-eslint/no-unsafe-call": "off",
-		"@typescript-eslint/no-unsafe-member-access": "off"
-	}
-}
