+++ src/MarkBlocked/modules/util/api.ts
@@ -1,4 +1,4 @@
-import * as OPTIONS from '../options.json';
+import * as OPTIONS from '../../options.json';
 import {initMwApi} from 'ext.gadget.Util';
 
 const api: mw.Api = initMwApi(`MarkBlocked/${OPTIONS.version}`);
